+++ bijection-avro/src/test/scala/com/twitter/bijection/avro/SpecificAvroCodecsSpecification.scala
@@ -1,6 +1,6 @@
 package com.twitter.bijection.avro
 
-import org.specs.Specification
+import org.scalatest._
 import com.twitter.bijection.{ Injection, BaseProperties }
 import org.apache.avro.Schema
 import avro.FiscalRecord
@@ -9,7 +9,7 @@ import avro.FiscalRecord
  * @author Muhammad Ashraf
  * @since 10/5/13
  */
-object SpecificAvroCodecsSpecification extends Specification with BaseProperties {
+object SpecificAvroCodecsSpecification extends WordSpec with Matchers with BaseProperties {
   val testSchema = new Schema.Parser().parse("""{
                                                    "type":"record",
                                                    "name":"FiscalRecord",
@@ -43,7 +43,47 @@ object SpecificAvroCodecsSpecification extends Specification with BaseProperties
       val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
       val bytes = Injection[FiscalRecord, Array[Byte]](testRecord)
       val attempt = Injection.invert[FiscalRecord, Array[Byte]](bytes)
-      attempt.get must_== testRecord
+      assert(attempt.get == testRecord)
+    }
+
+    "Round trip specific record using Specific Injection with Bzip2 compression" in {
+      implicit val specificInjection = SpecificAvroCodecs.withBzip2Compression[FiscalRecord]
+      val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
+      val bytes = Injection[FiscalRecord, Array[Byte]](testRecord)
+      val attempt = Injection.invert[FiscalRecord, Array[Byte]](bytes)
+      assert(attempt.get == testRecord)
+    }
+
+    "Round trip specific record using Specific Injection with Deflate compression (default compression level)" in {
+      implicit val specificInjection = SpecificAvroCodecs.withDeflateCompression[FiscalRecord]
+      val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
+      val bytes = Injection[FiscalRecord, Array[Byte]](testRecord)
+      val attempt = Injection.invert[FiscalRecord, Array[Byte]](bytes)
+      assert(attempt.get == testRecord)
+    }
+
+    "Round trip specific record using Specific Injection with Deflate compression (custom compression level)" in {
+      implicit val specificInjection = SpecificAvroCodecs.withDeflateCompression[FiscalRecord](9)
+      val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
+      val bytes = Injection[FiscalRecord, Array[Byte]](testRecord)
+      val attempt = Injection.invert[FiscalRecord, Array[Byte]](bytes)
+      assert(attempt.get == testRecord)
+    }
+
+    "Cannot create Specific Injection with Deflate compression if compression level is set too low" in {
+      an[IllegalArgumentException] should be thrownBy SpecificAvroCodecs.withDeflateCompression[FiscalRecord](0)
+    }
+
+    "Cannot create Specific Injection with Deflate compression if compression level is set too high" in {
+      an[IllegalArgumentException] should be thrownBy SpecificAvroCodecs.withDeflateCompression[FiscalRecord](10)
+    }
+
+    "Round trip specific record using Specific Injection with Snappy compression" in {
+      implicit val specificInjection = SpecificAvroCodecs.withSnappyCompression[FiscalRecord]
+      val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
+      val bytes = Injection[FiscalRecord, Array[Byte]](testRecord)
+      val attempt = Injection.invert[FiscalRecord, Array[Byte]](bytes)
+      assert(attempt.get == testRecord)
     }
 
     "Round trip specific record using Binary Injection" in {
@@ -51,7 +91,7 @@ object SpecificAvroCodecsSpecification extends Specification with BaseProperties
       val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
       val bytes = Injection[FiscalRecord, Array[Byte]](testRecord)
       val attempt = Injection.invert[FiscalRecord, Array[Byte]](bytes)
-      attempt.get must_== testRecord
+      assert(attempt.get == testRecord)
     }
 
     "Round trip specific record using Json Injection" in {
@@ -59,7 +99,7 @@ object SpecificAvroCodecsSpecification extends Specification with BaseProperties
       val testRecord = buildSpecificAvroRecord(("2012-01-01", 1, 12))
       val jsonString = Injection[FiscalRecord, String](testRecord)
       val attempt = Injection.invert[FiscalRecord, String](jsonString)
-      attempt.get must_== testRecord
+      assert(attempt.get == testRecord)
     }
   }