+++ bijection-core/src/main/scala/com/twitter/bijection/ClassInjection.scala
@@ -17,6 +17,7 @@ limitations under the License.
 package com.twitter.bijection
 
 import com.twitter.bijection.Inversion.attempt
+import scala.reflect.ClassTag
 
 /**
  *  Injection between Class objects and string.
@@ -33,8 +34,8 @@ class ClassInjection[T] extends AbstractInjection[Class[T], String] {
  * the inner items will not be correct. This is intended for experts.
  */
 object CastInjection {
-  def of[A, B >: A](implicit cmf: ClassManifest[A]): Injection[A, B] = new AbstractInjection[A, B] {
-    private val cls = cmf.erasure.asInstanceOf[Class[A]]
+  def of[A, B >: A](implicit cmf: ClassTag[A]): Injection[A, B] = new AbstractInjection[A, B] {
+    private val cls = cmf.runtimeClass.asInstanceOf[Class[A]]
     def apply(a: A) = a
     def invert(b: B) = attempt(b)(cls.cast(_))
   }