+++ bijection-core/src/test/scala/com/twitter/bijection/AsSyntax.scala
@@ -16,25 +16,23 @@ limitations under the License.
 
 package com.twitter.bijection
 
-import org.specs._
+import org.scalatest._
 import scala.util.Success
 
-class AsSyntax extends Specification {
-  noDetailedDiffs()
-
+class AsSyntax extends WordSpec with Matchers {
   import Conversion.asMethod
 
   "As syntax" should {
     "work on injections" in {
       val listAry = (1, 2).as[List[Array[Byte]]]
-      Injection.connect[(Int, Int), List[Array[Byte]]].invert(listAry) must be_==(Success(1 -> 2))
+      Injection.connect[(Int, Int), List[Array[Byte]]].invert(listAry) should be(Success(1 -> 2))
     }
     "work on bijections" in {
-      List(1, 2, 3).as[Vector[Int]] must be_==(Vector(1, 2, 3))
+      List(1, 2, 3).as[Vector[Int]] should be (Vector(1, 2, 3))
     }
     "work on functions" in {
       implicit def toS(i: Int): String = i.toString
-      23.as[String] must be_==("23")
+      23.as[String] should be ("23")
     }
   }