+++ bijection-jodatime/src/main/scala/com/twitter/bijection/jodatime/DateInjections.scala
@@ -1,7 +1,7 @@
 package com.twitter.bijection.jodatime
 
 import java.util.Date
-import org.joda.time.DateTime
+import org.joda.time.{ DateTime, LocalDate, LocalTime, YearMonth, MonthDay }
 import com.twitter.bijection.Inversion.attempt
 import com.twitter.bijection.{ Injection, InversionFailure, AbstractInjection }
 
@@ -24,5 +24,29 @@ trait DateInjections {
     new AbstractInjection[DateTime, Date] {
       def apply(d: DateTime) = d.toDate()
       override def invert(d: Date) = attempt(d)(new DateTime(_))
+    }
+
+  implicit val jodaLocalDate2String: Injection[LocalDate, String] =
+    new AbstractInjection[LocalDate, String] {
+      def apply(d: LocalDate) = d.toString
+      override def invert(s: String) = attempt(s)(LocalDate.parse(_))
+    }
+
+  implicit val jodaLocalTime2String: Injection[LocalTime, String] =
+    new AbstractInjection[LocalTime, String] {
+      def apply(d: LocalTime) = d.toString
+      override def invert(s: String) = attempt(s)(LocalTime.parse(_))
+    }
+
+  implicit val jodaYearMonth2String: Injection[YearMonth, String] =
+    new AbstractInjection[YearMonth, String] {
+      def apply(d: YearMonth) = d.toString
+      override def invert(s: String) = attempt(s)(YearMonth.parse(_))
+    }
+
+  implicit val jodaMonthDay2String: Injection[MonthDay, String] =
+    new AbstractInjection[MonthDay, String] {
+      def apply(d: MonthDay) = d.toString
+      override def invert(s: String) = attempt(s)(MonthDay.parse(_))
     }
 }