+++ bijection-jodatime/src/test/scala/com/twitter/bijection/jodatime/DateBijectionLaws.scala
@@ -1,28 +1,69 @@
 package com.twitter.bijection.jodatime
 
-import org.scalacheck.Properties
-import org.scalacheck.Gen._
-import org.scalacheck.Arbitrary
+import org.scalacheck.{ Arbitrary, Gen }
 import org.scalacheck.Prop._
+
+import org.scalatest.{ PropSpec, MustMatchers }
+import org.scalatest.prop.PropertyChecks
+
 import com.twitter.bijection.{ Bijection, BaseProperties, ImplicitBijection }
 import java.util.Date
-import org.joda.time.DateTime
+import org.joda.time.{ DateTime, LocalDate, LocalTime, YearMonth, MonthDay }
 import com.twitter.bijection._
 
-object DateBijectionsLaws extends Properties("DateBijections") with BaseProperties with DateBijections with DateInjections {
+class DateBijectionsLaws extends PropSpec with PropertyChecks with MustMatchers with BaseProperties with DateBijections with DateInjections {
 
   import Rep._
 
   implicit val strByte = arbitraryViaBijection[Date, String @@ Rep[Date]]
 
-  implicit val date = arbitraryViaFn { (dtime: Long) => new DateTime(dtime) }
+  /**
+   * If we generate times too close to the Max, timezone issues will push us over
+   */
+  case class Timestamp(ts: Long)
+  implicit val arbtimeStamp: Arbitrary[Timestamp] =
+    Arbitrary(Gen.choose(Long.MinValue / 4, Long.MaxValue / 4).map(Timestamp(_)))
 
-  property("Long <=> Joda") = isBijection[Long, DateTime]
+  implicit val date = arbitraryViaFn { (dtime: Timestamp) => new DateTime(dtime.ts) }
 
-  property("Date <=> Joda") = isBijection[Date, DateTime]
+  implicit val localDate = arbitraryViaFn { (dtime: Timestamp) => new LocalDate(dtime.ts) }
 
-  property("round trips Date -> String") = isLooseInjection[DateTime, String]
+  implicit val localTime = arbitraryViaFn { (dtime: Timestamp) => new LocalTime(dtime.ts) }
 
-  property("round trips Joda -> Date") = isLooseInjection[DateTime, Date]
+  implicit val yearMonth = arbitraryViaFn { (dtime: Timestamp) => new YearMonth(dtime.ts) }
+
+  implicit val monthDay = arbitraryViaFn { (dtime: Timestamp) => new MonthDay(dtime.ts) }
+
+  property("Long <=> Joda") {
+    isBijection[Long, DateTime]
+  }
+
+  property("Date <=> Joda") {
+    isBijection[Date, DateTime]
+  }
+
+  property("round trips Date -> String") {
+    isLooseInjection[DateTime, String]
+  }
+
+  property("round trips Joda -> Date") {
+    isLooseInjection[DateTime, Date]
+  }
+
+  property("round trips LocalDate -> String") {
+    isLooseInjection[LocalDate, String]
+  }
+
+  property("round trips LocalTime -> String") {
+    isLooseInjection[LocalTime, String]
+  }
+
+  property("round trips YearMonth -> String") {
+    isLooseInjection[YearMonth, String]
+  }
+
+  property("round trips MonthDay -> String") {
+    isLooseInjection[MonthDay, String]
+  }
 
 }